Refactoring: Notification Collection
