+++ app/frontend/containers/NotificationsContainer.js
@@ -3,8 +3,6 @@ import { push } from 'react-router-redux';
 import Notifications from '../ui/Notifications';
 import ApiClient from './ApiClient';
 import fetchNotifications from '../actions/fetchNotifications';
-import readNotification from '../actions/readNotification';
-import { sleep } from './Utils';
 import updatePageTitle from '../actions/updatePageTitle';
 import loadNotificationsStart from '../actions/loadNotificationsStart';
 import loadNotificationsEnd from '../actions/loadNotificationsEnd';
@@ -13,7 +11,6 @@ const mapStateToProps = state => {
   return {
     large: state.shared.large,
     notifications: state.shared.notifications,
-    unreadNotifications: state.shared.unreadNotifications,
     loadingNotifications: state.shared.loadingNotifications
   };
 };
@@ -30,28 +27,6 @@ const mapDispatchToProps = dispatch => {
         let notifications = await response.json();
         dispatch(fetchNotifications(notifications));
       }
-    },
-
-    readNotifications: async notifications => {
-      await sleep(5000);
-      const client = new ApiClient();
-      let unreadNotifications = notifications.filter(notification => {
-        return notification.read === false;
-      });
-      unreadNotifications.forEach(async notification => {
-        let response = await client.readNotification(notification.id);
-        if (response.ok) {
-          let notification = await response.json();
-          dispatch(readNotification(notification));
-        }
-        await sleep(3000);
-      });
-    },
-
-    handleNotificationClick: notification => {
-      dispatch(push(notification.click_action, {
-        previous: true
-      }));
     }
   };
 };
