+++ app/frontend/ui/Notifications.jsx
@@ -1,15 +1,8 @@
 import React, { Component } from 'react';
-import Typography from 'material-ui/Typography';
 import { CircularProgress } from 'material-ui/Progress';
-import List, {
-  ListItem,
-  ListItemText,
-  ListItemSecondaryAction
-} from 'material-ui/List';
-import Avatar from 'material-ui/Avatar';
 import Paper from 'material-ui/Paper';
-import moment from 'moment';
 import NoContentsContainer from '../containers/NoContentsContainer';
+import NotificationListContainer from '../containers/NotificationListContainer';
 
 const styles = {
   containerLarge: {
@@ -22,20 +15,12 @@ const styles = {
   progress: {
     textAlign: 'center',
     paddingTop: 40
-  },
-  notificationText: {
-    paddingRight: 32
-  },
-  secondaryAvatar: {
-    borderRadius: 0,
-    marginRight: 12
   }
 };
 
 class Notifications extends Component {
   componentWillMount() {
     this.props.handleMount();
-    this.props.readNotifications(this.props.notifications);
     gtag('config', process.env.GA_TRACKING_ID, {
       'page_path': '/notifications',
       'page_title': 'Notifications | Qoodish'
@@ -58,9 +43,7 @@ class Notifications extends Component {
     if (this.props.notifications.length > 0) {
       return (
         <Paper>
-          <List>
-            {this.renderNotifications(this.props.notifications)}
-          </List>
+          <NotificationListContainer notifications={this.props.notifications} />
         </Paper>
       );
     } else {
@@ -71,55 +54,6 @@ class Notifications extends Component {
         />
       );
     }
-  }
-
-  renderNotifications(notifications) {
-    return notifications.map(notification => (
-      <ListItem
-        onClick={() => {
-          this.props.handleNotificationClick(notification);
-        }}
-        key={notification.id}
-        button
-      >
-        <Avatar src={notification.notifier.profile_image_url} />
-        <ListItemText
-          primary={this.renderNotificationText(notification)}
-          secondary={
-            <div>{this.fromNow(notification)}</div>
-          }
-        />
-        {notification.notifiable.thumbnail_url && (
-          <ListItemSecondaryAction>
-            <Avatar src={notification.notifiable.thumbnail_url} style={styles.secondaryAvatar} />
-          </ListItemSecondaryAction>
-        )}
-      </ListItem>
-    ));
-  }
-
-  renderNotificationText(notification) {
-    if (notification.key == 'invited') {
-      return (
-        <div style={styles.notificationText}>
-          <b>{notification.notifier.name}</b> {notification.key} you to{' '}
-          {notification.notifiable.type}.
-        </div>
-      );
-    } else {
-      return (
-        <div style={styles.notificationText}>
-          <b>{notification.notifier.name}</b> {notification.key} your{' '}
-          {notification.notifiable.type}.
-        </div>
-      );
-    }
-  }
-
-  fromNow(notification) {
-    return moment(notification.created_at, 'YYYY-MM-DDThh:mm:ss.SSSZ')
-      .locale(window.currentLocale)
-      .fromNow();
   }
 
   renderProgress() {
